+++ src/component/jscoroutines.d.ts
@@ -1,22 +1,24 @@
 
-export declare function forEach(array: any[], fn: Function): void;
+
+
+export declare function forEach(collection: any, fn: Function): Generator<any, any, any>;
 
 /**
  * @returns array of elements matching the filter
  */
-export declare function filter(array: any[], fn: Function): any;
+export declare function filter(collection: any, fn: Function): Generator<any, any, any>;
 
 /**
  * @returns The result of processing the reduction function on all
  * of the items in the array
  */
-export declare function reduce(array: any[], fn: Function, initialValue?: any): any;
+export declare function reduce(collection: any, fn: Function, initialValue?: any): Generator<any, any, any>;
 
 /**
  * Concatenate two arrays into a new array
  * @returns the concatenated arrays
  */
-export declare function concat(array1: any[], array2: any[]): any[];
+export declare function concat(array1: any[], array2: any[]): Generator<any, any[], any>;
 
 /**
  * Appends one array to another
@@ -24,22 +26,22 @@ export declare function concat(array1: any[], array2: any[]): any[];
  * @param array2 - the source
  * @returns returns <code>array1</code>
  */
-export declare function append(array1: any[], array2: any[]): any[];
+export declare function append(array1: any[], array2: any[]): Generator<any, any[], any>;
 
 /**
  * @returns new array of mapped values
  */
-export declare function map(array: any[], fn: Function): any[];
+export declare function map(collection: any, fn: Function): Generator<any, any[], any>;
 
 /**
  * @returns the first matching value in the array or null
  */
-export declare function find(array: any[], fn: Function): any;
+export declare function find(collection: any, fn: Function): Generator<any, any, any>;
 
 /**
  * @returns Index of matching element or -1
  */
-export declare function findIndex(array: any[], fn: Function): number;
+export declare function findIndex(array: any[], fn: Function): Generator<any, number, any>;
 
 /**
  * @returns true if at least one item matched the filter
@@ -67,6 +69,130 @@ export declare function parseAsync(json: string): Promise<any>;
 
 
 /**
+ * Creates a generator for an array with the unique values from the
+ * input array, the routine is supplied with a
+ * function that determines on what the array should
+ * be made unique.
+ * @param {Array} array
+ * @param {Map} [fn] - the function to determine uniqueness, if
+ * omitted then the item itself is used
+ * @returns {Generator<*, [], *>}
+ */
+export declare function uniqueBy(array: [], fn: Function): Generator<any, [], any>
+
+/**
+ * Creates a generator for an object composed of keys generated from the results
+ * of running each element of collection thru then supplied function.
+ * The corresponding value of each key is an collection of the elements responsible
+ * for generating the key.
+ *
+ * @param {Array|Object} collection
+ * @param {Map} fn
+ * @returns {Generator<*, {}, *>} a generator for the new object
+ */
+export declare function groupBy(collection: [] | {}, fn: Function): Generator<any, {}, any>
+
+/**
+ * Creates an object composed of keys generated from the results
+ * of running each element of collection thru then supplied function.
+ * The corresponding value of each key is the last element responsible
+ * for generating the key.
+ *
+ * @param {Array|Object} collection
+ * @param {Map} fn
+ * @returns {Generator<*, {}, *>} a generator for the new object
+ */
+export declare function keyBy(collection: []|{}, fn: Function): Generator<any, {}, any>
+
+/**
+ * Returns a generator for the last index of an item in an array
+ * @param array - the array to scan
+ * @param value - the value to search for
+ * @returns {Generator<any, number, any>} a generator that returns the last index of the item or -1
+ */
+export declare function lastIndexOf(array: [], value: any): Generator<any, number, any>
+
+
+/**
+ * Returns a generator for an index of an item in an array
+ * @param array - the array to scan
+ * @param value - the value to search for
+ * @returns {Generator<any, number, any>} a generator that returns the index of the item or -1
+ */
+export declare function indexOf(array: [], value: any): Generator<any, number, any>
+
+/**
+ * Returns a generator returning true if an array includes a value
+ * @param array
+ * @param value
+ * @returns {Generator<*, boolean, *>}
+ */
+export declare function includes(array: [], value: any): Generator<any, boolean, any>;
+
+/**
+ * Creates a promise for an array with the unique values from the
+ * input array, the routine is supplied with a
+ * function that determines on what the array should
+ * be made unique.
+ * @param {Array} array
+ * @param {Map} [fn] - the function to determine uniqueness, if
+ * omitted then the item itself is used
+ * @returns {Promise<[]>}
+ */
+export declare function uniqueByAsync(array: [], fn: Function): Promise<[]>
+
+/**
+ * Creates a promise for an object composed of keys generated from the results
+ * of running each element of collection thru then supplied function.
+ * The corresponding value of each key is an collection of the elements responsible
+ * for generating the key.
+ *
+ * @param {Array|Object} collection
+ * @param {Map} fn
+ * @returns {Promise<{}>} a generator for the new object
+ */
+export declare function groupByAsync(collection: [] | {}, fn: Function): Promise<{}>
+
+/**
+ * Creates promise for an object composed of keys generated from the results
+ * of running each element of collection thru then supplied function.
+ * The corresponding value of each key is the last element responsible
+ * for generating the key.
+ *
+ * @param {Array|Object} collection
+ * @param {Map} fn
+ * @returns {Promise<{}>} a generator for the new object
+ */
+export declare function keyByAsync(collection: [] | {}, fn: Function): Promise<{}>
+
+/**
+ * Returns a promise for the last index of an item in an array
+ * @param array - the array to scan
+ * @param value - the value to search for
+ * @returns {Promise<Number>} a generator that returns the last index of the item or -1
+ */
+export declare function lastIndexOf(array: [], value: any): Promise<Number>
+
+
+/**
+ * Returns a promise for an index of an item in an array
+ * @param array - the array to scan
+ * @param value - the value to search for
+ * @returns {Promise<Number>} a generator that returns the index of the item or -1
+ */
+export declare function indexOf(array: [], value: any): Promise<number>
+
+/**
+ * Returns a promise for true if an array includes a value
+ * @param array
+ * @param value
+ * @returns {Promise<Boolean>}
+ */
+export declare function includes(array: [], value: any): Promise<Boolean>;
+
+
+
+/**
  * Sort an array (in place) by a sorting function
  * @example
  * async function process(data) {
@@ -75,32 +201,45 @@ export declare function parseAsync(json: string): Promise<any>;
  * @param array - The array to sort
  * @param sort - The method to sort the array
  * @returns a promise for the sorted array
+ * @example
+ *
+ * await sortAsync(myArray, v=>v.lastName)
  */
 export declare function sortAsync(array: any[], sort: Function): Promise<any[]>;
 
 /**
  * Finds an item in an array asynchronously
  * @returns promise for the item found or null if no match
+ * @example
+ *
+ * const firstToProcess = await findAsync(jobs, job=>job.done === false)
  */
-export declare function findAsync(array: any[], filter: Function): Promise<any | null>;
+export declare function findAsync(collection: []|{}, filter: Function): Promise<any | null>;
 
 /**
  * Finds an item index in an array asynchronously
  * @returns promise for the index of the first item to pass the filter or -1
+ * @example
+ * let firstItem = await findIndexAsync(jobs, job=>!!job.error)
  */
-export declare function findIndexAsync(array: any[], filter: Function): Promise<Number>;
+export declare function findIndexAsync(collection: []|{}, filter: Function): Promise<Number>;
 
 /**
  * Functions the contents of an array asynchronously
  * @returns promise for the mapped array
+ * @example
+ *
+ * const updated = await mapAsync(myItems, item=>({name: item.name, value: item.quantity * item.cost}))
  */
-export declare function mapAsync(array: any[], mapFn: Function): Promise<any[]>;
+export declare function mapAsync(collection: []|{}, mapFn: Function): Promise<[]|{}>;
 
 /**
  * Functions an array asynchronously
  * @returns promise for the filtered array
+ * @example
+ * const errorJobs = await filterAsync(
  */
-export declare function filterAsync(array: any[], filter: Function): Promise<any[]>;
+export declare function filterAsync(collection: []|{}, filter: Function): Promise<[]|{}>;
 
 /**
  * Performs a reduce on an array asynchronously
@@ -124,7 +263,7 @@ export declare function concatAsync(array1: any[], array2: any[]): Promise<any[]
  * Asynchronously loop over the elements of an array
  * @returns promise for the end of the operation
  */
-export declare function forEachAsync(array: any[], fn: Function): Promise<void>;
+export declare function forEachAsync(collection: []|{}, fn: Function): Promise<void>;
 
 /**
  * Asynchronously apply an array <code>some</code> operation
@@ -271,7 +410,7 @@ export declare function base64Decompress(base64Data: string): Promise<String>;
  * @param [timeout = 160 (ms)] - the number of milliseconds before the coroutine will run even if the system is not idle
  * @returns the result of the coroutine
  */
-export declare function run(coroutine: Function, loopWhileMsRemains?: number, timeout?: number): Promise<any>;
+export declare function run(coroutine: Function|Generator, loopWhileMsRemains?: number, timeout?: number): Promise<any>;
 
 /**
  * Start an animation coroutine, the animation will continue until
@@ -282,7 +421,7 @@ export declare function run(coroutine: Function, loopWhileMsRemains?: number, ti
  * @returns a value that will be returned to the caller
  * when the animation is complete.
  */
-export declare function update(coroutine: Function, ...params: any[]): Promise<any>;
+export declare function update(coroutine: Function|Generator, ...params: any[]): Promise<any>;
 
 /**
  * Starts an idle time coroutine using an async generator - this is NOT normally required
@@ -293,7 +432,7 @@ export declare function update(coroutine: Function, ...params: any[]): Promise<a
  * @param [timeout = 160 (ms)] - the number of milliseconds before the coroutine will run even if the system is not idle
  * @returns the result of the coroutine
  */
-export declare function runAsync(coroutine: Function, loopWhileMsRemains?: number, timeout?: number): Promise<any>;
+export declare function runAsync(coroutine: Function|Generator, loopWhileMsRemains?: number, timeout?: number): Promise<any>;
 
 /**
  * Wraps a normal function into a generator function
@@ -304,7 +443,7 @@ export declare function runAsync(coroutine: Function, loopWhileMsRemains?: numbe
  * @returns The wrapped yielding
  * version of the function passed
  */
-export declare function yielding(fn: (...params: any[]) => any, frequency?: number): Function;
+export declare function yielding(fn: (...params: any[]) => any, frequency?: number): (...params: any[]) => Generator;
 
 /**
  * Returns a function that will execute the passed
@@ -315,7 +454,7 @@ export declare function yielding(fn: (...params: any[]) => any, frequency?: numb
  * @returns a function that can be called to execute the coroutine
  * and return its result on completion
  */
-export declare function wrapAsPromise(coroutine: Function): Function;
+export declare function wrapAsPromise(coroutine: ()=>Generator): Function;
 
 export declare function pipe(...functions: Array<Function>|Array<Array<Function>>) : any
 export declare function tap(fn: Function) : Function;